added default runtime to interpreter crate
