+++ blockchain/state_manager/Cargo.toml
@@ -12,10 +12,8 @@ db = { path = "../../node/db/" }
 encoding = { package = "forest_encoding", path = "../../encoding/" }
 num-bigint = { path = "../../utils/bigint", package = "forest_bigint" }
 state_tree = { path = "../../vm/state_tree/" }
-default_runtime = { path = "../../vm/default_runtime/" }
 blockstore = { package = "ipld_blockstore", path = "../../ipld/blockstore/" }
 forest_blocks = { path = "../../blockchain/blocks" }
 thiserror = "1.0"
 interpreter = { path = "../../vm/interpreter/" }
-runtime = { path = "../../vm/runtime/" }
 ipld_amt = { path = "../../ipld/amt/" }
