+++ blockchain/state_manager/src/lib.rs
@@ -9,13 +9,11 @@ use address::{Address, Protocol};
 use blockstore::BlockStore;
 use blockstore::BufferedBlockStore;
 use cid::Cid;
-use default_runtime::resolve_to_key_addr;
 use encoding::de::DeserializeOwned;
 use forest_blocks::FullTipset;
-use interpreter::VM;
+use interpreter::{resolve_to_key_addr, DefaultSyscalls, VM};
 use ipld_amt::Amt;
 use num_bigint::BigUint;
-use runtime::DefaultSyscalls;
 use state_tree::StateTree;
 use std::error::Error as StdError;
 use std::sync::Arc;
