--- vm/default_runtime/Cargo.toml
@@ -1,27 +1,0 @@
-[package]
-name = "default_runtime"
-version = "0.1.0"
-authors = ["ChainSafe Systems <info@chainsafe.io>"]
-edition = "2018"
-
-[dependencies]
-address = { package = "forest_address", path = "../address" }
-actor = { path = "../actor" }
-message = { package = "forest_message", path = "../message" }
-runtime = { path = "../runtime" }
-state_tree = { path = "../state_tree" }
-clock = { path = "../../node/clock" }
-vm = { path = "../" }
-ipld_blockstore = { path = "../../ipld/blockstore" }
-num-bigint = { path = "../../utils/bigint", package = "forest_bigint" }
-forest_encoding = { path = "../../encoding" }
-cid = { package = "forest_cid", path = "../../ipld/cid" }
-crypto = { path = "../../crypto" }
-num-traits = "0.2.11"
-byteorder = "1.3.4"
-db = { path = "../../node/db" }
-
-[dev-dependencies]
-state_tree = { path = "../state_tree" }
-db = { path = "../../node/db" }
-ipld_hamt = { path = "../../ipld/hamt" }
